Add Twisted support to the benchmark too
